+++ summarizer/models/transformer.py
@@ -148,7 +148,7 @@ class TransformerModel(Model):
                 target /= target.max()
 
                 if self.hps.use_cuda:
-                    seq, target = seq.float().cuda(), target.float().cuda()
+                    seq, target = seq.cuda(), target.cuda()
 
                 y = self.model(seq)
 
@@ -185,7 +185,7 @@ class TransformerModel(Model):
                 seq = torch.from_numpy(seq).unsqueeze(0)
 
                 if self.hps.use_cuda:
-                    seq = seq.float().cuda()
+                    seq = seq.cuda()
 
                 y = self.model(seq)
                 summary[key] = y[0].detach().cpu().numpy()
