+++ summarizer/models/vasnet.py
@@ -193,7 +193,7 @@ class VASNetModel(Model):
                 target /= target.max()
 
                 if self.hps.use_cuda:
-                    seq, target = seq.float().cuda(), target.float().cuda()
+                    seq, target = seq.cuda(), target.cuda()
 
                 y = self.model(seq)
 
@@ -230,7 +230,7 @@ class VASNetModel(Model):
                 seq = torch.from_numpy(seq).unsqueeze(0)
 
                 if self.hps.use_cuda:
-                    seq = seq.float().cuda()
+                    seq = seq.cuda()
 
                 y = self.model(seq)
                 summary[key] = y[0].detach().cpu().numpy()
